Aufsplitten der Routen
