+++ turnierseite/app.py
@@ -60,12 +60,19 @@ def create_app():
 
     # import and register blueprints
     from turnierseite.core.routes import core
-    from turnierseite.turnier.routes import turnier
     from turnierseite.user_handeling.routes import user_handeling
+    #from turnierseite.turnier.routes import turnier
+    from turnierseite.turnier.routes.turnier import turnier
+    from turnierseite.turnier.routes.gruppe import gruppe
+    from turnierseite.turnier.routes.team import team
+    from turnierseite.turnier.routes.spielplan import spielplan
 
     app.register_blueprint(core, url_prefix='/')
-    app.register_blueprint(turnier, url_prefix='/turnier')
     app.register_blueprint(user_handeling, url_prefix='/user_handeling', bcrypt=bcrypt)
+    app.register_blueprint(turnier, url_prefix='/turnier')
+    app.register_blueprint(gruppe, url_prefix='/gruppe')
+    app.register_blueprint(team, url_prefix='/team')
+    app.register_blueprint(spielplan, url_prefix='/spielplan')
 
     csrf.exempt(core)
     csrf.exempt(user_handeling)
